pyproject.toml
[project]
name = "spreadsheet-to-pdf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fpdf2>=2.8.3",
    "openpyxl>=3.1.5",
    "pandas>=2.3.0",
    "pandas-stubs>=2.3.0.250703",
    "types-fpdf2>=2.8.3.20250516",
    "weasyprint>=65.1",
]
